requirements.txt
httpx[http2]>=0.28.0
python-dotenv>=1.0.0